requirements.txt
beautifulsoup4==4.12.3
fastapi==0.115.0
lxml==5.3.0
opencc-python-reimplemented==0.1.7
python-multipart==0.0.9
uvicorn==0.30.6
